+++ peacecorps/peacecorps/templates/donations/review_payment.jinja
@@ -26,12 +26,11 @@
 
 
     <form action="{{oci_servlet_url}}" method="POST">
-        <input type="hidden" name="agency_memo" value="{{agency_memo}}"/>
         <input type="hidden" name="agency_id" value="{{agency_id}}"/>
-        <input type="hidden" name="agency_tracking_id" value="{{tracking_id}}"/>
+        <input type="hidden" name="agency_tracking_id"
+                             value="{{agency_tracking_id}}"/>
         <input type="hidden" name="app_name" value="{{app_name}}"/>
-        <input type="hidden" name="senddata" value="Enter Payment Information"/>
-        <input type="submit" value="Enter Payment Information" />
+        <input type="submit" name="senddata" value="Enter Payment Information"/>
     </form>
 </body>
 </html>
